Song Analyser Building Github - just segmented/requirements.txt
requests
python-dotenv
lyricsgenius
beautifulsoup
